+++ experimental_sdk/.rubocop.yml
@@ -1,0 +1,22 @@
+AllCops:
+  TargetRubyVersion: "2.6.0"
+
+Lint/UnusedMethodArgument:
+  Enabled: false
+Metrics/AbcSize:
+  Max: 30
+Metrics/LineLength:
+  Enabled: false
+Metrics/MethodLength:
+  Max: 50
+Metrics/PerceivedComplexity:
+  Max: 30
+Metrics/CyclomaticComplexity:
+  Max: 20
+Metrics/ParameterLists:
+  Enabled: false
+Naming/FileName:
+  Exclude:
+    - "lib/opentelemetry-experimental-sdk.rb"
+Style/ModuleFunction:
+  Enabled: false
