+++ experimental_sdk/Gemfile
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+source 'https://rubygems.org'
+
+gemspec
+
+gem 'opentelemetry-api', path: '../api'
+gem 'opentelemetry-sdk', path: '../sdk'
+gem 'opentelemetry-test-helpers', path: '../test_helpers'
+
+group :test, :development do
+  gem 'pry'
+  gem 'pry-byebug' unless RUBY_ENGINE == 'jruby'
+end
