+++ experimental_sdk/lib/opentelemetry-experimental-sdk.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+require 'opentelemetry/sdk'
+require 'opentelemetry/sdk/experimental'
