+++ experimental_sdk/lib/opentelemetry/sdk/experimental/version.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+module OpenTelemetry
+  module SDK
+    module Experimental
+      # Current OpenTelemetry experimental sdk version
+      VERSION = '0.0.1'
+    end
+  end
+end
